% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_constraint.R
\name{add_distance_constraint}
\alias{add_distance_constraint}
\title{Apply distance constraint}
\usage{
add_distance_constraint(in.dat, distance.threshold)
}
\arguments{
\item{in.dat}{potential points to sample from}

\item{distance.threshold}{minimum distance between points}
}
\value{
sf object with randomly sampled points
}
\description{
select final camera locations based on distance constraints
}
