% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_camera_locations.R
\name{generate_cam_locations}
\alias{generate_cam_locations}
\title{Generate Camera Locations}
\usage{
generate_cam_locations(poly.dat, area.threshold, num.per.poly)
}
\arguments{
\item{num.per.poly}{}
}
\description{
Generate a set of possible camera locations across camera grid
}
