% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_union.R
\name{GIS_union}
\alias{GIS_union}
\title{GIS union}
\usage{
GIS_union(a, b)
}
\arguments{
\item{a}{the first sf object}

\item{b}{the second sf object}
}
\value{
an sf object that unions sf a and b
}
\description{
perform a real GIS union operation, similar to QGIS or ArcGIS
}
