% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make Grid}
\usage{
make_grid(
  dat,
  cellsize,
  org.x = NULL,
  org.y = NULL,
  num.cols = NULL,
  area.threshold,
  plot.proportion = TRUE
)
}
\arguments{
\item{area.threshold}{}
}
\description{
generate a grid over a polygon
}
